Refactor: move Schedule to its own module
