+++ app/src/main/java/ch/epfl/sweng/eventmanager/inject/ActivityBuilder.java
@@ -2,6 +2,7 @@ package ch.epfl.sweng.eventmanager.inject;
 
 import ch.epfl.sweng.eventmanager.ui.eventSelector.EventPickingModule;
 import ch.epfl.sweng.eventmanager.ui.eventShowcase.EventShowcaseModule;
+import ch.epfl.sweng.eventmanager.ui.schedule.ScheduleModule;
 import dagger.Module;
 import dagger.android.AndroidInjectionModule;
 
@@ -10,7 +11,7 @@ import dagger.android.AndroidInjectionModule;
  * When creating a new activity, its module should be registered here.
  * @author Louis Vialar
  */
-@Module(includes = {AndroidInjectionModule.class, EventPickingModule.class, EventShowcaseModule.class})
+@Module(includes = {AndroidInjectionModule.class, EventPickingModule.class, EventShowcaseModule.class, ScheduleModule.class})
 public abstract class ActivityBuilder {
 
     /*
