+++ app/src/main/java/ch/epfl/sweng/eventmanager/ui/eventShowcase/EventActivity.java
@@ -10,6 +10,7 @@ import android.widget.TextView;
 
 import ch.epfl.sweng.eventmanager.R;
 import ch.epfl.sweng.eventmanager.ui.eventSelector.EventPickingActivity;
+import ch.epfl.sweng.eventmanager.ui.schedule.ScheduleActivity;
 import ch.epfl.sweng.eventmanager.viewmodel.ViewModelFactory;
 import dagger.android.AndroidInjection;
