+++ app/src/main/java/ch/epfl/sweng/eventmanager/ui/eventShowcase/EventShowcaseModule.java
@@ -1,14 +1,11 @@
 package ch.epfl.sweng.eventmanager.ui.eventShowcase;
 
-import android.app.Activity;
 import android.arch.lifecycle.ViewModel;
 import android.arch.lifecycle.ViewModelProvider;
 import ch.epfl.sweng.eventmanager.viewmodel.ViewModelFactory;
 import ch.epfl.sweng.eventmanager.viewmodel.ViewModelKey;
 import dagger.Binds;
 import dagger.Module;
-import dagger.android.ActivityKey;
-import dagger.android.AndroidInjector;
 import dagger.android.ContributesAndroidInjector;
 import dagger.multibindings.IntoMap;
 
@@ -25,15 +22,11 @@ public abstract class EventShowcaseModule {
     @ViewModelKey(EventShowcaseModel.class)
     abstract ViewModel provideEventListModel(EventShowcaseModel eventListModel);
 
-
     @ContributesAndroidInjector
     abstract EventActivity contributeEventActivityInjector();
 
     @ContributesAndroidInjector
     abstract MapActivity contributeMapActivityInjector();
-
-    @ContributesAndroidInjector
-    abstract ScheduleActivity contributeScheduleActivityInjector();
 
     @ContributesAndroidInjector
     abstract TicketActivity contributeTicketActivityInjector();
