+++ app/src/main/java/ch/epfl/sweng/eventmanager/ui/schedule/ScheduleModule.java
@@ -1,0 +1,13 @@
+package ch.epfl.sweng.eventmanager.ui.schedule;
+
+import dagger.Module;
+import dagger.android.ContributesAndroidInjector;
+
+/**
+ * @author Louis Vialar
+ */
+@Module
+public abstract class ScheduleModule {
+    @ContributesAndroidInjector
+    abstract ScheduleActivity contributeScheduleActivityInjector();
+}
