+++ app/src/main/java/ch/epfl/sweng/eventmanager/ui/schedule/TimeLineModel.java
@@ -1,4 +1,4 @@
-package ch.epfl.sweng.eventmanager.ui.eventShowcase.scheduleUtils;
+package ch.epfl.sweng.eventmanager.ui.schedule;
 
 import android.os.Parcel;
 import android.os.Parcelable;
