+++ app/src/main/res/layout/activity_schedule.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.eventShowcase.ScheduleActivity">
+    tools:context=".ui.schedule.ScheduleActivity">
 
 
     <include layout="@layout/content_timeline"></include>
