+++ app/src/main/res/layout/content_timeline.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="ch.epfl.sweng.eventmanager.ui.eventShowcase.ScheduleActivity"
+    tools:context="ch.epfl.sweng.eventmanager.ui.schedule.ScheduleActivity"
     tools:showIn="@layout/activity_schedule">
 
     <android.support.v7.widget.RecyclerView
